requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
google-api-python-client==2.108.0
google-generativeai==0.3.2
python-dotenv==1.0.0
urllib3==2.0.7
lxml==4.9.3
aiohttp==3.9.1
validators==0.22.0
tldextract==5.1.1 